requirement.txt
streamlit
panda
numpy
requests
altair
